mm/madvise, fs/binfmt_elf: scale range alignment and batching to process page size
Update madvise_set_anon_name() alignment check and length rounding to use
mm_pte_aligned() and mm_pte_align(). Scale swapin_walk_pmd_entry(),
madvise_cold_or_pageout_pte_range(), and madvise_free_pte_range() address
steps and TLB page size updates to mm_pte_size().

In fs/binfmt_elf.c, cast ELF_PAGESTART() alignment mask to unsigned long.

+++ fs/binfmt_elf.c
@@ -89,7 +89,7 @@ static inline unsigned long elf_min_align(struct mm_struct *mm)
 #define ELF_CORE_EFLAGS	0
 #endif
 
-#define ELF_PAGESTART(mm, _v)	((_v) & ~(int)(ELF_MIN_ALIGN(mm) - 1))
+#define ELF_PAGESTART(mm, _v)	((_v) & ~(unsigned long)(ELF_MIN_ALIGN(mm) - 1))
 #define ELF_PAGEOFFSET(mm, _v)	((_v) & (ELF_MIN_ALIGN(mm) - 1))
 #define ELF_PAGEALIGN(mm, _v)	roundup((_v), ELF_MIN_ALIGN(mm))
